requirements.txt
flet
python-dotenv
supabase-py